Select a Telegram channel from a folder where multiple archives are stored

Select a Telegram channel from a folder where multiple archives are stored

Usage

tp_select(
  path = NULL,
  channel_name = NULL,
  channel_id = NULL,
  only_latest = TRUE
)

Arguments

path

Path to a folder where Telegram archives in json format have previously been stored.

channel_name

Defaults to NULL. Name of one ore more channels, must match exactly. Be aware that, in principle, different channels can have exactly the same name or can change name, channel_id is generally safer.

channel_id

Defaults to NULL. Numeric identifier of one ore more channels, must match exactly.

Value

A vector of paths to json files corresponding to the relevant Telegram channel.