Set the channel username for locally stored Telegram channels

Channel username is necessary to re-create links to channels and posts. Unfortunately, it cannot be retrieved from the export json files themselves and needs to be added manually.

Usage

tp_set_username(
  channel_name = NULL,
  channel_id = NULL,
  channel_username = NULL,
  path = NULL
)

Arguments

channel_username

Username of the Telegram channel, either as link (e.g. "https://t.me/username"), as username ("@username"), or as plain text ("username"). Either way, the username is stored consistently internally as plain text.

path

Path to a folder where a tp_usernames.csv file is expected to be located. This will usually be the same folder where Telegram archives are stored.

Value

A data frame with one row and three columns (channel_id, channel_name, channel_username) if relevant data has previously been stored. Otherwise, returns NULL.

Details

Information is cached locally in a file called tp_usernames.csv located in the path folder.

If no channel_username is given, the user will be requested to add it interactively.

Examples

tp_set_username(
  channel_name = "BBC News | Русская служба",
  channel_username = "bbcrussian"
)
#> Error in tp_select(path = path, channel_name = channel_name): `path` must be a folder, and /home/runner/telegramparser isn't.